requirements.txt
pyspark==3.4.1
pyarrow==12.0.1
pandas==2.0.3
sqlalchemy==2.0.19
python-dotenv==1.0.0
